refactor: separate router business logic feat/#27
