+++ services/product-service/app/services/__init__.py
@@ -1,0 +1,1 @@
+"""product-service business service package."""
